feat(gql-status-object): Introduce GqlNotification (#1667)
BREAKING CHANGE: this update separates the new GQL Status Object hierarchy from the legacy `Notification` type by introducing a new `GqlNotification` interface that extends the `GqlStatusObject`.

The update is based on the feedback received on the GQL Status Object feature, which is currently in preview. This should make it easier to understand the new GQL Status Objects and migrate from using the legacy `Notification` to the new types.

Eventually, the new hierarchy is expected to replace the legacy `Notification` that is expected to be deprecated and phased out at a later date.