feat(unsupportedType): add support for Bolt Unsupported Type (#1691)
This update adds support for a new `TypeSystem#UNSUPPORTED()` type. The unsupported type may occur when a new type is introduced in the Neo4j server and the driver is connected over an older Bolt Protocol version that does not support this type. A new `UnsupportedType` object provides information about the unsupported type and the minimum Bolt version needed to support it. However, the `UnsupportedType` object itself requires at least Bolt Protocol 6.0.